==> scripts/30_calculate_fbi_and_pop_indicators.Rmd <==
---
title : "Assessing the effect of climate change on a bioassessment scheme: Lessons from a long-term river fish monitoring in France"
subtitle: "Calculation of population and FBI-related indicators"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  bookdown::html_document2:
    toc: true
    toc_float: true
#  bookdown::word_document2
always_allow_html: true
params:
  refnet_start: 2013
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Setup

## Load packages and functions

```{r}
# if necessary
# devtools::install_github("pascalirz/aspe")
library(aspe)
library(tidyverse)

# install custom functions
source("../R/functions.R")
```

## Load data

```{r}
rdata_tables <- misc_nom_dernier_fichier(
  repertoire = "../../../../../projets/ASPE/raw_data/do_not_delete_fbi_trend_version_do_not_delete",
  pattern = "^tables")
# load it
load(rdata_tables)


load(file = "../processed_data/20_selected_data.RData")
```

Data file : `r str_replace(rdata_tables, "../../../../projets/ASPE/raw_data/rdata/", "")`

> NB The analysis is carried out considering the REFNET starts in `r params$refnet_start`.

# Fish-based Index

## Aggregated index

Here we add a color palette to use as background in plots. A distinction in the quality classes threshold has recently been introduced depending on site altitude, but here we stick to the genuine thresholds, ie we remove the thresholds corresponding to the sites over 500m in altitude.

```{r}
fbi_classes <- classe_ipr %>%
  aspe::ip_completer_classes_couleur() %>% 
  filter(cli_altitude_min != 500 | is.na(cli_altitude_min))
```

Assembly of the dataset

```{r}
fbi <- surveys %>% 
  mef_ajouter_ipr() %>% # add FBI score
  mef_ajouter_metriques() %>% # add FBI metrics scores
  # add hydrographic districts
  left_join(y = point_prelevement %>% 
              select(pop_id, unh_id = pop_unh_id)) %>% 
  left_join(y = ref_unite_hydrographique %>% 
              select(unh_id, unh_libelle)) %>% 
  select(-unh_id)
```

Add the reference and/or non-reference status of the survey

> NB this duplicates the surveys contributing to more than one

```{r}
fbi <- fbi %>% 
  left_join(points_ref_status_2_cats %>% 
              select(-n_ope))

# remove the surveys on "reference sites" before REFNET was set up
fbi <- fbi %>% 
  filter(!(network == "REFNET" & annee < params$refnet_start))
```

### For the complete REPNET and REFNET 

Calculation of quantiles for the `geom_ribbon()` plot.

```{r}
fbi_quartiles <- fbi %>% 
  group_by(annee, network) %>% 
  summarise(p25 = quantile(ipr, 0.25, na.rm = TRUE),
            p50 = quantile(ipr, 0.50, na.rm = TRUE),
            p75 = quantile(ipr, 0.75, na.rm = TRUE)) %>% 
  ungroup()

fbi_pc_quality_class <- fbi %>% 
  group_by(annee, network) %>% 
  summarise(pc_very_good = sum(ipr < 5) / n(),
            pc_good_or_better = sum(ipr < 16) / n(),
            pc_mediocre_or_better = sum(ipr < 25) / n(),
            pc_bad_or_better = sum(ipr < 36) / n()) %>% 
  ungroup()

fbi_quartiles <- fbi_quartiles %>% 
  left_join(y = fbi_pc_quality_class)
```

Calculate percentage of surveys with a FBI score indicating "Good ecological condition" for each network.

```{r}
fbi_pc_good <- fbi %>% 
  group_by(network) %>% 
  summarise(pc_good_or_better = sum(ipr < 16) / n()) %>% 
  ungroup()
```

### For the REPNET and REFNET excluding shared sites

```{r}
fbi_quartiles_exclusive <- fbi %>% 
  filter(pop_id %in% pop_id_exclusive_network) %>% 
  group_by(annee, network) %>% 
  summarise(p25 = quantile(ipr, 0.25, na.rm = TRUE),
            p50 = quantile(ipr, 0.50, na.rm = TRUE),
            p75 = quantile(ipr, 0.75, na.rm = TRUE)) %>% 
  ungroup()
```

Calculate percentage of surveys with a FBI score indicating "Good ecological condition" for each network.

```{r}
fbi_pc_good_exclusive <- fbi %>% 
  filter(pop_id %in% pop_id_exclusive_network) %>% 
  group_by(network) %>% 
  summarise(pc_good_or_better = sum(ipr < 16) / n()) %>% 
  ungroup()
```

## FBI metrics

Build a dataframe with the FBI metrics, organised in 'long' format.

```{r, eval=TRUE}
fbi_metrics <- fbi %>% 
  select(sta_id:annee, ner:dti, unh_libelle, network) %>% 
  # put into "long format" for the faceted plot
  pivot_longer(cols = ner:dti,
               names_to = "metric",
               values_to = "value") %>% 
  # translate metrics abbreviations into ENG labels
  mutate(metric = fct_recode(metric,
    "dens_invertivorous_indiv" = "dii",
    "dens_omnivorous_indiv" = "dio",
    "dens_tolerant_indiv" = "dit",
    "dens_total_indiv" = "dti",
    "nb_lithophilic_sp" = "nel",
    "nb_rheophilic_sp" = "ner",
    "nb_total_sp" = "nte"),
         # for a nicer plot
         metric = fct_rev(metric))

fbi_metrics %>% 
  head() %>% 
  mutate(across(c(sta_id:ope_id, annee), as.character)) %>% 
  flextable::flextable()
```

Calculate the annual median of each metric across sites

```{r}
fbi_metrics_median <- fbi_metrics %>% 
  group_by(annee, metric, network) %>% 
  summarise(p50 = median(value, na.rm = TRUE)) %>% 
  ungroup()
```

# Populations trends

## Species occurrence probability

This information is in the Aspe database and obtained using the function `aspe::mef_ajouter_proba_presence_ipr()`.

The surveys carried out before `r params$refnet_start` and tagged "REFNET" are removed from the dataset.

```{r}
# Assembly of a temp dataframe
sp_proba_init <- fbi %>% 
  select(sta_id:annee, network) %>% 
  distinct() %>% 
  filter(!(network == "REFNET" & annee < params$refnet_start)) %>% 
  # add probabilities of occurrence from dataframe "probabilite_presence_ipr"
  mef_ajouter_proba_presence_ipr() %>% 
  rename(sp_proba_occur = ppi_valeur_probabilite, # proba occurrence
         obs_nb_indiv = ppi_param_effectif # observed catches
  ) %>% 
  # add species scientific name
  left_join(y = ref_espece %>% 
              select(esp_id, # species id
                     esp_nom_latin # species scientific name
              )) %>% 
  # recode some inacurate species names
  latin_recoding(var_latin_name = esp_nom_latin)

sp_fbi_conversion <- sp_proba_init %>% 
  left_join(ref_espece) %>% 
  select(esp_code_alternatif, esp_nom_latin) %>% 
  distinct() %>% 
  mutate(esp_code_alternatif = case_when(
    str_detect(esp_nom_latin, "Cottus") ~ "CHA",
    str_detect(esp_nom_latin, "Squalius") ~ "CHE",
    str_detect(esp_nom_latin, "Leuciscus") ~ "VAN",
    str_detect(esp_nom_latin, "Esox") ~ "BRO",
    str_detect(esp_nom_latin, "Carassius") ~ "CAX",
    str_detect(esp_nom_latin, "Barbatula") ~ "LOF",
    str_detect(esp_nom_latin, "Thymallus") ~ "OBR",
    str_detect(esp_nom_latin, "Phoxinus") ~ "PHX",
    str_detect(esp_nom_latin, "Gobio") ~ "GOX",
    TRUE ~ esp_code_alternatif
  ))
```

The species are considered as predicted present if, and only if, their modeled probability of occurrence exceeds 0.5 (cf. Oberdorff et al. (2002) Development and validation of a fish‐based index for the assessment of 'river health' in France. Freshwater Biology, 47(9), 1720-1734).

```{r}
sp_proba <- sp_proba_init %>%
  mutate(sp_pred_presence = sp_proba_occur >= 0.5, # threshold between pres and abs
         sp_obs_presence = obs_nb_indiv > 0) %>%
  select(ope_id,
         annee,
         network,
         esp_nom_latin,
         sp_pred_presence,
         sp_obs_presence)
```

Here we distinguish the cases as in a confusion matrix (good model prediction, false positive and false negative).

```{r}
confusion1 <- sp_proba %>%
  mutate(
    case = case_when(
      sp_pred_presence & sp_obs_presence ~ "pred_pres_obs_pres",
      !sp_pred_presence & sp_obs_presence ~ "pred_abs_obs_pres",
      sp_pred_presence & !sp_obs_presence ~ "pred_pres_obs_abs",
      !sp_pred_presence & !sp_obs_presence ~ "pred_abs_obs_abs"
    )
  )

# nb surveys by year
n_surveys_by_year <- sp_proba %>%
  group_by(annee, network) %>%
  summarise(n_surveys = n_distinct(ope_id))

# count of confusion cases
confusion <- confusion1 %>%
  group_by(esp_nom_latin, annee, network, case) %>%
  tally() %>%
  ungroup()

# completing missing combinations
confusion <- confusion %>%
  select(esp_nom_latin, annee, case, network, n) %>%
  complete(esp_nom_latin, annee, case, fill = list(n = 0)) %>%
  left_join(y = n_surveys_by_year) %>%
  pivot_wider(names_from = case,
              values_from = n) %>%
  # calculation of percentages
  mutate(
    pc_pred_abs_obs_pres = pred_abs_obs_pres / n_surveys,
    pc_pred_pres_obs_abs = pred_pres_obs_abs / n_surveys,
    pc_pred_pres_obs_pres = pred_pres_obs_pres / n_surveys,
    pc_pred_abs_obs_abs = pred_abs_obs_abs / n_surveys
  )

confusion %>% 
  sample_n(10) %>% 
  mutate(across(annee, as.character)) %>% 
  flextable::flextable()
```

## Species contributions to the FBI metrics

This data is downloaded from ['Système d'évaluation de l'état des eaux' (SEEE) website](https://seee.eaufrance.fr/), following \> Outil d'évaluation \> IPR \> Documentation \> acces algorithmes \> IPR_params_Especes_Contributives.csv

```{r}
sp_guilds <- read_csv2(file = "../raw_data/IPR_params_Especes_Contributives.csv") %>%
  rename(esp_code_alternatif = Taxon) %>%
  left_join(y = ref_espece %>%
              select(esp_code_alternatif,
                     esp_nom_latin),
            by = "esp_code_alternatif") %>% 
  left_join(y = ref_espece %>%
            select(esp_code_alternatif,
                     esp_nom_latin_fbi = esp_nom_latin),
            by = c("REGROUP" = "esp_code_alternatif")) %>% 
  # pool species according to the FBI taxonomy
  latin_recoding(var_latin_name = esp_nom_latin_fbi) %>% 
  latin_recoding(var_latin_name = esp_nom_latin) %>% 
  mutate(esp_nom_latin_fbi = ifelse(
    test = is.na(esp_nom_latin_fbi),
    yes = esp_nom_latin,
    no = esp_nom_latin_fbi
  )) %>% 
  select(esp_code_alternatif,
    esp_nom_latin,
         esp_nom_latin_fbi,
         NER:DIT) %>% 
  # Englishisize metrics names and levels
  rename(
    "dens_invertivorous_indiv" = "DII",
    "dens_omnivorous_indiv" = "DIO",
    "dens_tolerant_indiv" = "DIT",
    "dens_total_indiv" = "DTI",
    "nb_lithophilic_sp" = "NEL",
    "nb_rheophilic_sp" = "NER",
    "nb_total_sp" = "NTE"
  ) %>% 
  pivot_longer(cols = nb_rheophilic_sp:dens_tolerant_indiv) %>% 
  mutate(value = case_when(value == "O" ~ "Yes",
                           value == "N" ~ "No",
                           value == "R" ~ "Pooled",
                           TRUE ~ NA_character_)) %>% 
  pivot_wider(names_from = name,
              values_from = value)

flextable::flextable(head(sp_guilds))
```

## Species densities

For each species, the densities (in number of individuals per 1000 m²) are first calculated for each survey, then aggregated by year (median). The absences, with zero densities, are not considered to avoid redundancy (or colinearity) with occurrence. 

```{r}
sp_densities <- sp_proba_init %>% 
  mef_ajouter_surf_calc() %>% 
  filter(obs_nb_indiv > 0) %>% 
  mutate(density = 1000 * obs_nb_indiv / ope_surface_calculee) %>% 
  group_by(annee, esp_nom_latin, network) %>% 
  summarise(density = median(density))
```

## Species occurency rate across sites

For each species and each year, this indicator is the ratio of the number of sites where the species was observed by the number of sites surveyed.

```{r}
sp_occupancy <- sp_proba_init %>% 
  filter(obs_nb_indiv > 0) %>% 
  group_by(annee, esp_nom_latin, network) %>% 
  summarise(n_occur = n_distinct(ope_id)) %>% 
  ungroup() %>% 
  left_join(n_surveys_by_year) %>% 
  mutate(occurency_rate = n_occur / n_surveys)
```

## Assembly of population indicators

```{r}
sp_occ <- sp_occupancy %>% 
  select(-n_occur, -n_surveys) %>% 
  rename(value = occurency_rate) %>% 
  mutate(indicator = "occurency_rate")
  
names(sp_occupancy)
names(sp_densities)
```

# Saving

```{r}
save(fbi,
     fbi_quartiles,
     fbi_pc_good,
     pop_id_exclusive_network,
     fbi_quartiles_exclusive,
     fbi_pc_good_exclusive,
     fbi_metrics,
     points_ref_status_2_cats,
     fbi_metrics_median,
     fbi_classes,
     points_geo,
     points,
     sp_proba,
     confusion1,
     confusion,
     classe_ipr,
     sp_occupancy,
     sp_densities,
     ref_espece,
     ref_unite_hydrographique,
     sp_fbi_conversion,
     file = "../processed_data/30_indicators.RData")
```
